+++ src/Journeys/JourneyNode/JourneyAddToDigestNode/Type.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Courier\Journeys\JourneyNode\JourneyAddToDigestNode;
+
+enum Type: string
+{
+    case ADD_TO_DIGEST = 'add-to-digest';
+}